:art: move SemanticVersionFormatter to Formatters folder
