+++ UI/Bootstrapper.cs
@@ -1,8 +1,8 @@
 ﻿using System.IO.Abstractions;
 using Autofac;
 using MemoryPack;
-using MuseDashModToolsUI.Converters;
 using MuseDashModToolsUI.Extensions.MarkupExtensions;
+using MuseDashModToolsUI.Formatters;
 using MuseDashModToolsUI.ViewModels;
 using MuseDashModToolsUI.ViewModels.Dialogs;
 using MuseDashModToolsUI.ViewModels.Pages;
